feat: implemented the welcome image
