+++ src/actions/onGuildMemberAdd.ts
@@ -1,6 +1,6 @@
 import { GuildMember } from 'discord.js';
 
-import { onWelcome } from '../actionHandlers/welcome';
+import { onWelcome } from '../actionHandlers/onWelcome';
 
 export const onGuildMemberAdd = (member: GuildMember) => {
   onWelcome(member);
